Improve logging from webhook
When an error occurs in the webhook (either with itself or due to
invalid input), it returns a 4xx/5xx and some error text. For security
reasons, Caddy reduces all of these results to a simple 400/503. Since
aiohttp only logs the response code, and Caddy only logs admin problems,
this information was lost.

We don't really do anything with access logs, so just enable them only
for the errors.