Fix top-level brochure site
Formerly, `_static` was not used by the brochure site, and so was
redirected to `/stable/_static` as part of removing old top-level URLs.

At some point, either an upgrade to Sphinx or to the theme put files
back into `_static`. This was fine when both used the same versions of
things, but https://github.com/matplotlib/mpl-brochure-site/pull/103
updated themes in the brochure site while the stable docs remained on an
older version.

Since the brochure site is not coupled with the stable docs, remove the
redirects for these directories.

Fixes matplotlib/mpl-brochure-site#109